sw_services: xilpuf: Add glitch detection check for Status
This patch adds glitch detection check for Status in XPuf_Regeneration.
It also adds typecasting for Status in XSECURE_STATUS_CHK_GLITCH_DETECT.

Acked-by: Kundanala, Praveen Teja <praveen.teja.kundanala@amd.com>

+++ lib/bsp/standalone/src/common/xil_util.h
@@ -198,7 +198,7 @@ extern "C" {
 #define XSECURE_STATUS_CHK_GLITCH_DETECT(Status)   \
 	{ \
 		if (Status == XST_SUCCESS) { \
-			Status = XST_GLITCH_ERROR; \
+			Status = (int)XST_GLITCH_ERROR; \
 		} \
 	}
